Moved quoteType to objects folder
