+++ cms/schemaTypes/index.ts
@@ -3,6 +3,6 @@ import {articleType} from './articleType'
 import {eventType} from './eventType'
 import {roleType} from './roleType'
 import customImage from './objects/customImage'
-import {quoteType} from './quoteType'
+import {quoteType} from './objects/quoteType'
 
 export const schemaTypes = [articleType, eventType, frontpage, roleType, customImage, quoteType]
